# Speed up `mapped ngrams()`