--- a/layout_evaluation/src/ngram_mapper/bigram_mapper.rs
+++ b/layout_evaluation/src/ngram_mapper/bigram_mapper.rs
@@ -5,13 +5,14 @@
 //! for secondary bigrams. Not only those that lead to same-hand bigrams.
 
 use super::{common::*, on_demand_ngram_mapper::SplitModifiersConfig};
-use super::{BigramIndices, TrigramIndices};
+use super::{BigramIndices, BigramIndicesVec, TrigramIndices};
 
 use crate::ngrams::Bigrams;
 
 use keyboard_layout::layout::{LayerKey, Layout};
 
-use rustc_hash::{FxHashMap, FxHashSet};
+use ahash::AHashMap;
+use rustc_hash::FxHashSet;
 use serde::Deserialize;
 
 /// Configuration parameters for process of increasing the weight of common bigrams.
@@ -43,30 +44,21 @@ impl Default for IncreaseCommonBigramsConfig {
 
 /// Increase the weight of bigrams that already have a weight exceeding a threshold even further.
 pub fn increase_common_bigrams(
-    bigram_keys: &BigramIndices,
+    bigram_keys: &mut BigramIndices,
     config: &IncreaseCommonBigramsConfig,
-) -> BigramIndices {
+) {
     if !config.enabled {
-        return bigram_keys.to_vec();
+        return;
     }
 
-    // here we need to collect all mapped trigrams per key in order to successfully increase weights in next step
-    let mut m = FxHashMap::default();
-    bigram_keys.iter().for_each(|(k, w)| {
-        *m.entry(*k).or_insert(0.0) += *w;
-    });
-
-    let total_weight: f64 = m.values().sum();
+    let total_weight: f64 = bigram_keys.values().sum();
     let critical_point = config.critical_fraction * total_weight;
 
-    m.iter_mut().for_each(|(_c, weight)| {
-        let mut new_weight = *weight;
+    bigram_keys.values_mut().for_each(|weight| {
         if *weight > critical_point && total_weight > config.total_weight_threshold {
-            new_weight += (new_weight - critical_point) * (config.factor - 1.0);
+            *weight += (*weight - critical_point) * (config.factor - 1.0);
         }
-        *weight = new_weight;
     });
-    m.into_iter().collect()
 }
 
 /// Configuration parameters for adding secondary bigrams from trigrams.
@@ -105,7 +97,6 @@ pub fn add_secondary_bigrams_from_trigrams(
     }
 
     // there are many duplicates in the secondary bigrams -> using a hashmap is cheaper
-    let mut m = FxHashMap::with_capacity_and_hasher(trigram_keys.len(), Default::default());
     trigram_keys
         .iter()
         .map(|((idx1, idx2, idx3), w)| {
@@ -133,17 +124,17 @@ pub fn add_secondary_bigrams_from_trigrams(
                     config.factor_handswitch
                 };
 
-                *m.entry((*idx1, *idx3)).or_insert(0.0) += *weight * factor;
+                bigram_keys.insert_or_add_weight((*idx1, *idx3), weight * factor);
             },
         );
-    bigram_keys.extend(m);
 }
 
-fn layerkey_indices(
+/// Turns the [`Bigrams`]'s characters into their indices, returning a [`BigramIndicesVec`].
+fn mapped_bigrams(
     bigrams: &Bigrams,
     layout: &Layout,
     exclude_line_breaks: bool,
-) -> (BigramIndices, f64) {
+) -> (BigramIndicesVec, f64) {
     let mut not_found_weight = 0.0;
     let mut bigram_keys = Vec::with_capacity(bigrams.grams.len());
 
@@ -197,14 +188,17 @@ impl OnDemandBigramMapper {
         exclude_line_breaks: bool,
     ) -> (BigramIndices, f64, f64) {
         // println!("Before split: {:?}", self.bigrams.grams.get(&('l', 'r')));
-        let (mut bigram_keys, not_found_weight) =
-            layerkey_indices(&self.bigrams, layout, exclude_line_breaks);
+        let (bigram_keys_vec, not_found_weight) =
+            mapped_bigrams(&self.bigrams, layout, exclude_line_breaks);
 
-        if self.split_modifiers.enabled {
-            bigram_keys = self.split_bigram_modifiers(&bigram_keys, layout);
-        }
+        let found_weight: f64 = bigram_keys_vec.iter().map(|(_, w)| w).sum();
 
-        let found_weight = bigram_keys.iter().map(|(_, w)| w).sum();
+        let bigram_keys = if self.split_modifiers.enabled {
+            self.split_bigram_modifiers(bigram_keys_vec, layout)
+        } else {
+            bigram_keys_vec.into_iter().collect()
+        };
+
         // bigram_keys
         //     .iter()
         //     .filter(|((c1, c2), _)| c1.symbol == 'l' && c2.symbol == 'r')
@@ -231,41 +225,41 @@ impl OnDemandBigramMapper {
     ///
     /// Each bigram of higher-layer symbols will transform into a series of bigrams with permutations of
     /// the involved base-keys and modifers. However, the base-key will always be after its modifier.
-    fn split_bigram_modifiers(&self, bigrams: &BigramIndices, layout: &Layout) -> BigramIndices {
-        let mut bigram_keys = Vec::with_capacity(2 * bigrams.len());
+    fn split_bigram_modifiers(&self, bigrams: BigramIndicesVec, layout: &Layout) -> BigramIndices {
+        let mut bigram_w_map = AHashMap::with_capacity(bigrams.len() / 3);
 
-        bigrams.iter().for_each(|((k1, k2), w)| {
-            let (base1, mods1) = layout.resolve_modifiers(k1);
-            let (base2, mods2) = layout.resolve_modifiers(k2);
+        bigrams.into_iter().for_each(|((k1, k2), w)| {
+            let (base1, mods1) = layout.resolve_modifiers(&k1);
+            let (base2, mods2) = layout.resolve_modifiers(&k2);
 
-            bigram_keys.push(((base1, base2), *w));
+            bigram_w_map.insert_or_add_weight((base1, base2), w);
 
             mods1.iter().for_each(|mod1| {
                 // mix mods of k1 with base of k2
-                bigram_keys.push(((*mod1, base2), *w * 0.5));
+                bigram_w_map.insert_or_add_weight((*mod1, base2), w * 0.5);
 
                 // mix mods of k1 and k2
                 mods2.iter().for_each(|mod2| {
                     if mod1 != mod2 {
-                        bigram_keys.push(((*mod1, *mod2), *w));
+                        bigram_w_map.insert_or_add_weight((*mod1, *mod2), w);
                     }
                 });
             });
 
             mods2.iter().for_each(|mod2| {
                 // mix mods of k2 with base of k1
-                bigram_keys.push(((base1, *mod2), *w * 2.0));
+                bigram_w_map.insert_or_add_weight((base1, *mod2), w * 2.0);
             });
 
             // same key mods
-            TakeTwoLayerKey::new(base1, &mods1, *w, self.split_modifiers.same_key_mod_factor)
+            TakeTwoLayerKey::new(base1, &mods1, w, self.split_modifiers.same_key_mod_factor)
                 .for_each(|(e, w)| {
-                    bigram_keys.push((e, w));
+                    bigram_w_map.insert_or_add_weight(e, w);
                 });
 
-            TakeTwoLayerKey::new(base2, &mods2, *w, self.split_modifiers.same_key_mod_factor)
+            TakeTwoLayerKey::new(base2, &mods2, w, self.split_modifiers.same_key_mod_factor)
                 .for_each(|(e, w)| {
-                    bigram_keys.push((e, w));
+                    bigram_w_map.insert_or_add_weight(e, w);
                 });
 
             // log::debug!(
@@ -283,6 +277,6 @@ impl OnDemandBigramMapper {
             // );
         });
 
-        bigram_keys
+        bigram_w_map
     }
 }