--- a/layout_evaluation/src/ngram_mapper/bigram_mapper.rs
+++ b/layout_evaluation/src/ngram_mapper/bigram_mapper.rs
@@ -224,7 +224,22 @@ impl OnDemandBigramMapper {
     ) -> Vec<((&'s LayerKey, &'s LayerKey), f64)> {
         bigrams
             .iter()
-            .map(|((k1, k2), w)| ((layout.get_layerkey(k1), layout.get_layerkey(k2)), *w))
+            .filter_map(|((idx1, idx2), w)| {
+                let k1 = layout.get_layerkey(idx1);
+
+                // If the same modifier appears consecutively, it is usually "hold" instead of repeatedly pressed
+                // --> remove
+                match k1.is_modifier && idx1 == idx2 {
+                    false => Some((
+                        (
+                            k1,                        // LayerKey 1
+                            layout.get_layerkey(idx2), // LayerKey 2
+                        ),
+                        *w,
+                    )),
+                    true => None,
+                }
+            })
             .collect()
     }