@Glitchy-Tozier 2022-04-02
One last – possibly unnecessary idea I have is the following:
- We find out which characters are `fixed_keys:` or part of `fixed_layers: [2, 3]`.
- We split our corpus into two halves:
  - The ngrams that _only_ contain fixed characters
  - The ngrams that contain any number of unfixed chars
- Then, we prepare the fixed chars only once
- Now every time we call `map_unigrams`, we only map unfixed chars and finally add them to the fixed ones.

The fixed chars & fixed layers include the following characters: 
```
…_[]^!<>=&ſ\/{}*?()-:@#$|~`+%"';1234567890¡¿:−ªº№·£¤\t\n/*−°§ℓ»«$€„“”₁₂₃♀♂⚥ϰ⟨⟩₀‑¬∨∧⊥∡∥→∞∝    ∅­
```
However, when using regex and searching for trigrams which contain those characters
```regex
 […_\[\]\^!<>=&ſ\\/\{\}\*\?\(\)-:@#\$\|~`\+%"';1234567890¡¿:−ªº№·£¤\n\t/\*−°§ℓ»«\$€„“”₁₂₃♀♂⚥ϰ⟨⟩₀‑¬∨∧⊥∡∥→∞∝    ∅­]{3}\n
```
, we find only 16_000 of the 290_000 Trigrams contain only fixed chars. Thus, this idea probably doesn't help much in terms of performance.