@dariogoetz 2022-04-02
I have mixed feelings about this pull request.

1. On one hand, I see that using a `HashMap` instead of the `Vec` for the `*GramIndices` saves some conversions. But I am a little uncomfortable with then having both structures there. Is it possible to get rid of the `*GramIndicesVec` version and only convert them into a `Vec` in the last step, when the values are being placed into the `MappedNgrams`? Then we would only keep the `HashMap` version and thus save the conversions.

2. Also I see that there are now multiple instances where a `AHashMap` is used in favor of the `FxHashMap`. I would prefer being consistent here and use `AHashMap` everywhere (if performance does not degrade too much) while completely getting rid of the `FxHashMap`. This then also removes the `rustc_hash` dependency. Did you check what the performance impact would be using only `AHashMap`s?

3. Finally, I see your reasoning for implementing the `PartialEq` manually. However, I do not like having the `index` inside the `LayerKey`. I understand that it is only meant to be a "unique" value. But using the `LayerKeyIndex` is very misleading as it encodes information "from the outside". How much time does using the "automatic" implementation of `PartialEq` actually cost?