@Glitchy-Tozier 2022-04-03
Hi :)

> Your commit

The `/ 3` in `let mut bigram_w_map = AHashMap::with_capacity(bigrams.len() / 3);` was intentional. Due to the fact that we do not fill the hashmap with any duplicates (and due to the fact that we only fill it with Layer1-letters), it's usually smaller than 1/3rd or ever 1/4th or 1/5th of `bigrams.len()`. Thus I added the `/ 3` because it seems to be the most efficient thing to do.
I'd prefer keeping the `/ 3`, but if you know of scenarios I forgot, I'm happy to remove the `/ 3`.

> 1. On one hand, I see that using a `HashMap` instead of the `Vec` for the `*GramIndices` saves some conversions. But I am a little uncomfortable with then having both structures there. Is it possible to get rid of the `*GramIndicesVec` version and only convert them into a `Vec` in the last step, when the values are being placed into the `MappedNgrams`? Then we would only keep the `HashMap` version and thus save the conversions.

I share your feeling that this is ugly.
In my tests this actually had a performance impact, and was the main thing that saved time. This might be due to one of two reasons:
1. Inserting Items into a HashMap might be slower than pushing them into a Vec (which would make sense as additional hashing-computations are performed)
2. Iterating through a HashMap might be slower (no idea how reasonable this idea is)

That being said:
1. my tests might have been imperfect
2. we could just remove the type but still use a Vec to return the values from that one function.


> 2. Also I see that there are now multiple instances where a `AHashMap` is used in favor of the `FxHashMap`. I would prefer being consistent here and use `AHashMap` everywhere (if performance does not degrade too much) while completely getting rid of the `FxHashMap`. This then also removes the `rustc_hash` dependency.

I started using `AHashMap` due to [their comparison to other hashers](https://github.com/tkaitchuck/aHash/blob/master/compare/readme.md). Accordingly, I did my best to consintently use one or the other using the following rule:
- FxHash for primitive keys
- AHash for more complex keys such as Tuples or Strings.

> Did you check what the performance impact would be using only `AHashMap`s?

I did not and it's probably negligible. We can use `AHashMap` everywhere, if you prefer. :)
If we decide on removing FxHash…:
- Let's wait until the two PRs are gone, to avoid constant merge-conflicts
- It's probably most efficient to replace the really small `FxHashSets` (such as `exclude_rows` or `initial_pause_indicators`) with Vecs

> Finally, I see your reasoning for implementing the PartialEq manually. However, I do not like having the index inside the LayerKey. I understand that it is only meant to be a "unique" value. But using the LayerKeyIndex is very misleading as it encodes information "from the outside". How much time does using the "automatic" implementation of PartialEq actually cost?

> How much time does using the "automatic" implementation of PartialEq actually cost?

Not much. I assume this is because most of the time the first fields are not equal to each other, and thus the "comparison-train" is stopped preemptively.

> But using the LayerKeyIndex is very misleading as it encodes information "from the outside".

I don't quite understand this part. Is it problematic because it references a "higher order concept", such as the layout? `LayerKeyIndex = "This is my position in the Layout"`?
In that case, my proposal is the following:
Rename `pub index: LayerKeyIndex` into `pub id: u16`. This way, we do not imply any relation to the layout (and thus are not misleading), but still are able to do efficient comparisons. What do you think?