@Glitchy-Tozier 2022-04-03
_Actually_ … there's no need at all to fill either a `Vec` or a `HashMap` in `fn map_trigrams`. It's redundant. We use the value only once (to iterate over it), thus we can simply return an Iterator.

This skips
1. Pushing > 200_000 trigrams into a vec
2. turning that vec back into an iterable.

I've got a fun & consise idea on how to implement this, if you give me the go :)